+++ app.js
@@ -1,7 +1,9 @@
-
 const express = require('express');
 const exphbs = require('express-handlebars');
 const bodyParser = require('body-parser');
+const multer = require('multer');
+const cloudinaryStorage = require('multer-storage-cloudinary');
+const cloudinary = require('cloudinary').v2;
 
 const app = express();
 const port = 3000;
@@ -13,15 +15,34 @@ app.set('view engine', 'handlebars');
 // Middleware to parse form data
 app.use(bodyParser.urlencoded({ extended: true }));
 
+// Configure Cloudinary
+cloudinary.config({
+    cloud_name: 'dwqsanfkj',
+    api_key: '456521387282719',
+    api_secret: 'oT-Ks7m6_gCMoY3kek8I171diy8'
+});
+
+// Configure multer with Cloudinary storage
+const storage = cloudinaryStorage({
+    cloudinary: cloudinary,
+    folder: 'uploads', // optional, specify the folder for uploaded files
+    allowedFormats: ['jpg', 'jpeg', 'png'],
+    filename: function (req, file, cb) {
+        cb(null, file.originalname); // use the original file name as the Cloudinary public_id
+    }
+});
+
+const parser = multer({ storage: storage });
+
 // Serve the HTML form
 app.get('/', (req, res) => {
     res.render('form');
 });
 
-// Handle form submission
-app.post('/userprofile', (req, res) => {
+// Handle form submission with file upload
+app.post('/userprofile', parser.single('image'), (req, res) => {
     const { title, author, ingredients, steps } = req.body;
-    const imageFile = req.files ? req.files.image.name : 'Not selected'; // Assuming you are using file upload middleware
+    const imageFile = req.file ? req.file.url : 'Not selected'; // Assuming you are using multer with multer-storage-cloudinary
 
     // Render the userprofile.handlebars template with the submitted data
     res.render('userprofile', { title, author, ingredients, steps, imageFile });
