+++ controllers/api/index.js
@@ -1,0 +1,10 @@
+const router = require('express').Router();
+const postRoutes = require('./postRoutes');
+const userRoutes = require('./userRoutes');
+
+router.use('/posts', postRoutes);
+
+router.use('/users', userRoutes);
+
+module.exports = router;
+
