+++ controllers/index.js
@@ -1,0 +1,9 @@
+const router = require('express').Router();
+
+const apiRoutes = require('./api');
+const homeRoutes = require('./homeRoutes');
+
+router.use('/', homeRoutes);
+router.use('/api', apiRoutes);
+
+module.exports = router;
