+++ seeds/imageData.json
@@ -1,0 +1,10 @@
+[
+    {
+        "multimedia_url":"https://res.cloudinary.com/dwqsanfkj/image/upload/v1705971361/tjdwqurbr5gasn8uioxe.png",
+        "user_id":1
+    },
+    {
+        "multimedia_url":"https://res.cloudinary.com/dwqsanfkj/image/upload/v1705971466/k7nwdlgmjkfxsnaqgis9.jpg",
+        "user_id":2
+    }
+]
